Add example 1.5 from the text and update .gitignore
The first example of flex and bison being used together,
we build a simple "calculator" to interpret basic arithmetic
expressions from input.

Building has already become a bit more complex, so we now
use a Makefile for this purpose.

The .gitignore file is updated to ignore all yacc output
files.